Fix locale preloading, fixes #9, #10.
Preloads.js is now a plugin that preloads the data for the specified locales,
and returns a hash with that data.

Also created load! helper plugin, to load the data for a single specified locale.
It makes the preloads.js code more understandable and facilitates future enhancement
when a locale can have a variable number of objects, rather than always three.